+++ .gitignore
@@ -1,0 +1,108 @@
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# C extensions
+*.so
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+pip-wheel-metadata/
+share/python-wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+
+# PyInstaller
+# Usually these files are written by a python script from a template
+# before PyInstaller builds the exe, so as long as the template exists
+# it should be fine
+# MANIFEST.in
+
+# Translations
+/*.mo
+/*.pot
+
+# Django stuff:
+# *.log
+# local_settings.py
+# db.sqlite3
+
+# Flask stuff:
+# instance/
+.webassets-cache
+
+# Scrapy stuff:
+.scrapy
+
+# Sphinx documentation
+docs/_build/
+
+# PyBuilder
+.pybuilder/
+target/
+
+# Rope project
+.ropeproject
+
+# mkdocs documentation
+/site/
+
+# pytest cache or coverage
+.pytest_cache/
+.coverage
+#*cover
+
+# celery beat schedule file
+celerybeat-schedule
+
+# SageMath parsed files
+*.sage.py
+
+# Environments
+.env
+.venv
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
+
+# Spyder project settings
+.spyderproject
+
+# Rope, Jedi
+.ropeproject/
+.history
+
+# mypy
+.mypy_cache/
+tmp/mypy/
+
+# IPython notebooks
+.ipynb_checkpoints
+
+# pyc files
+**/*.pyc
+
+# environments
+.env/
+venv/
+
+# System-related files
+.DS_Store
+
+# card_data_cache.json
+card_data_cache.json
