+++ pyproject.toml
@@ -1,0 +1,25 @@
+[project]
+name = "glimmer-grabber"
+version = "0.1.0"
+description = "Tooling for digitizing Lorcana Trading Cards."
+authors = [{ name = "Your Name", email = "your.email@example.com" }]  # Replace with actual author info if available
+license = {text = "MIT"}
+requires-python = ">=3.7"
+dependencies = [
+    "numpy",
+    "opencv-python",
+    "scikit-image",
+    "requests",
+    "ultralytics",
+]
+
+[project.scripts]
+glimmer-grabber = "glimmer_grabber.cli:main"  # CLI Entry Point
+
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+packages = ["glimmer_grabber"]
+package-dir = {"" = "src"}
