+++ src/glimmer_grabber/grayscale_converter.py
