+++ src/glimmer_grabber/illumination_normalizer.py
