+++ src/glimmer_grabber/noise_reducer.py
